Add initialization part to Main and first GET index page request
Main reason for such method is driven from Hibernate + Spark + Testing conceptions:

"taskDao" and "gson" are defined "static" with getters, so that we can use them in tests.

"Main.main" accepts two arguments: PORT and hibernate configuration file. That is used purely in testing, to load testing database with another PORT number.

GET request to index page uses WebConstants class, that is proved to be useful, to keep track of the constants used here and there.

This "Main" class is tested for GET method for index page