Add getMaxId method and implementation
in both TaskDao and TaskDaoImpl. This can be used in testing, because we re-use database, we have to know how many tasks are there, and keep track of the added one, to know that task was removed.

May be it is Unit Testing violation, but I still find it useful